Initial work to split header types
