+++ sdk/core/Azure.Core/src/Request.cs
@@ -121,7 +121,7 @@ namespace Azure.Core
         /// TBD.
         /// </summary>
         /// <returns></returns>
-        protected override PipelineMessageHeaders GetHeadersCore()
+        protected override PipelineRequestHeaders GetHeadersCore()
             => new AzureCoreMessageHeaders(Headers);
 
         #endregion
@@ -182,7 +182,7 @@ namespace Azure.Core
         /// <summary>
         /// Backwards adapter to MessageHeaders to implement GetHeadersCore
         /// </summary>
-        private sealed class AzureCoreMessageHeaders : PipelineMessageHeaders
+        private sealed class AzureCoreMessageHeaders : PipelineRequestHeaders
         {
             /// <summary>
             /// Headers on the Azure.Core.Request type to adapt to.
