+++ sdk/core/System.ClientModel/tests/TestFramework/Mocks/MockResponseHeaders.cs
@@ -1,0 +1,33 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.ClientModel.Primitives;
+using System.Collections.Generic;
+
+namespace ClientModel.Tests.Mocks;
+
+public class MockResponseHeaders : PipelineResponseHeaders
+{
+    private readonly Dictionary<string, string> _headers;
+
+    public MockResponseHeaders()
+    {
+        _headers = new Dictionary<string, string>();
+    }
+
+    public override IEnumerator<KeyValuePair<string, string>> GetEnumerator()
+    {
+        throw new NotImplementedException();
+    }
+
+    public override bool TryGetValue(string name, out string? value)
+    {
+        return _headers.TryGetValue(name, out value);
+    }
+
+    public override bool TryGetValues(string name, out IEnumerable<string>? values)
+    {
+        throw new NotImplementedException();
+    }
+}
