+++ sdk/core/System.ClientModel/tests/internal/Message/PipelineResponseHeadersTests.cs
@@ -12,45 +12,13 @@ namespace System.ClientModel.Tests.Message;
 public class PipelineResponseHeadersTests
 {
     [Test]
-    public void AddThrowsNotSupportedException()
-    {
-        HttpResponseMessage httpResponse = new HttpResponseMessage();
-        HttpContent content = new ByteArrayContent(BinaryData.FromString("Content").ToArray());
-
-        HttpClientPipelineResponseHeaders headers = new(httpResponse, content);
-
-        Assert.Throws<NotSupportedException>(() => headers.Add("Mock-Header", "Mock value"));
-    }
-
-    [Test]
-    public void SetThrowsNotSupportedException()
-    {
-        HttpResponseMessage httpResponse = new HttpResponseMessage();
-        HttpContent content = new ByteArrayContent(BinaryData.FromString("Content").ToArray());
-
-        HttpClientPipelineResponseHeaders headers = new(httpResponse, content);
-
-        Assert.Throws<NotSupportedException>(() => headers.Set("Mock-Header", "Mock value"));
-    }
-
-    [Test]
-    public void RemoveThrowsNotSupportedException()
-    {
-        HttpResponseMessage httpResponse = new HttpResponseMessage();
-        HttpContent content = new ByteArrayContent(BinaryData.FromString("Content").ToArray());
-        HttpClientPipelineResponseHeaders headers = new(httpResponse, content);
-
-        Assert.Throws<NotSupportedException>(() => headers.Remove("Mock-Header"));
-    }
-
-    [Test]
     public void CanGetSingleValueAsString()
     {
         HttpResponseMessage httpResponse = new HttpResponseMessage();
         httpResponse.Headers.Add("Mock-Header", "Mock value");
         HttpContent content = new ByteArrayContent(BinaryData.FromString("Content").ToArray());
 
-        HttpClientPipelineResponseHeaders headers = new(httpResponse, content);
+        HttpClientResponseHeaders headers = new(httpResponse, content);
 
         Assert.IsTrue(headers.TryGetValue("Mock-Header", out string? value));
         Assert.AreEqual("Mock value", value);
@@ -64,7 +32,7 @@ public class PipelineResponseHeadersTests
         httpResponse.Headers.Add("Mock-Header", "Value 2");
         HttpContent content = new ByteArrayContent(BinaryData.FromString("Content").ToArray());
 
-        HttpClientPipelineResponseHeaders headers = new(httpResponse, content);
+        HttpClientResponseHeaders headers = new(httpResponse, content);
 
         Assert.IsTrue(headers.TryGetValue("Mock-Header", out string? value));
         Assert.AreEqual("Value 1,Value 2", value);
@@ -77,7 +45,7 @@ public class PipelineResponseHeadersTests
         httpResponse.Headers.Add("Mock-Header", "Mock value");
         HttpContent content = new ByteArrayContent(BinaryData.FromString("Content").ToArray());
 
-        HttpClientPipelineResponseHeaders headers = new(httpResponse, content);
+        HttpClientResponseHeaders headers = new(httpResponse, content);
 
         Assert.IsTrue(headers.TryGetValues("Mock-Header", out IEnumerable<string>? values));
         Assert.AreEqual(1, values!.Count());
@@ -92,7 +60,7 @@ public class PipelineResponseHeadersTests
         httpResponse.Headers.Add("Mock-Header", "Value 2");
         HttpContent content = new ByteArrayContent(BinaryData.FromString("Content").ToArray());
 
-        HttpClientPipelineResponseHeaders headers = new(httpResponse, content);
+        HttpClientResponseHeaders headers = new(httpResponse, content);
         Assert.IsTrue(headers.TryGetValues("Mock-Header", out IEnumerable<string>? values));
         Assert.AreEqual(2, values!.Count());
         Assert.AreEqual("Value 1", values!.ElementAt(0));
@@ -106,7 +74,7 @@ public class PipelineResponseHeadersTests
         httpResponse.Headers.Add("Mock-Header", "Mock value");
         HttpContent content = new ByteArrayContent(BinaryData.FromString("Content").ToArray());
 
-        HttpClientPipelineResponseHeaders headers = new(httpResponse, content);
+        HttpClientResponseHeaders headers = new(httpResponse, content);
 
         Assert.IsTrue(headers.TryGetValues("MOCK-HEADER", out IEnumerable<string>? values));
         Assert.AreEqual(1, values!.Count());
